Fix: spacing parameters float instead of int.

+++ src/PhpPresentation/Reader/PowerPoint2007.php
@@ -1129,15 +1129,15 @@ class PowerPoint2007 implements ReaderInterface
             $oElementLineSpacingPercent = $xmlReader->getElement('a:lnSpc/a:spcPct', $oSubElement);
             if ($oElementLineSpacingPercent instanceof DOMElement) {
                 $oParagraph->setLineSpacingMode(Paragraph::LINE_SPACING_MODE_PERCENT);
-                $oParagraph->setLineSpacing((int)($oElementLineSpacingPercent->getAttribute('val') / 1000));
+                $oParagraph->setLineSpacing($oElementLineSpacingPercent->getAttribute('val') / 1000);
             }
             $oElementSpacingBefore = $xmlReader->getElement('a:spcBef/a:spcPts', $oSubElement);
             if ($oElementSpacingBefore instanceof DOMElement) {
-                $oParagraph->setSpacingBefore((int)($oElementSpacingBefore->getAttribute('val') / 100));
+                $oParagraph->setSpacingBefore($oElementSpacingBefore->getAttribute('val') / 100);
             }
             $oElementSpacingAfter = $xmlReader->getElement('a:spcAft/a:spcPts', $oSubElement);
             if ($oElementSpacingAfter instanceof DOMElement) {
-                $oParagraph->setSpacingAfter((int)($oElementSpacingAfter->getAttribute('val') / 100));
+                $oParagraph->setSpacingAfter($oElementSpacingAfter->getAttribute('val') / 100);
             }
 
             $oParagraph->getBulletStyle()->setBulletType(Bullet::TYPE_NONE);
